Move Code Into Main File, Attempt To Add Test For :text-center
